+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## 18-09-2017
+### Added
+- asynctesting: nestedpromises
+- test:watch
+
 ## 08-09-2017
 ### Added
 - functional: either.